refactor(tool-router): split search and policy into separate modules (fixes #129)
